Update Speech to match new sidekit api

+++ dnntoolkit.py
@@ -1249,7 +1249,8 @@ class Speech():
     def logmel(signal, fs, n_filters=40, n_ceps=13,
             win=0.025, shift=0.01,
     		delta1=True, delta2=True, energy=True,
-    		normalize=True, clean=True, vad=True, returnVAD=False):
+    		normalize=True, clean=True,
+            vad=True, returnVAD=False):
         import sidekit
 
         if len(signal.shape) > 1:
@@ -1260,7 +1261,7 @@ class Speech():
         # n_filters = 40 # The number of mel filter bands
         f_min = 0. # The minimal frequency of the filter bank
         f_max = fs / 2
-        nwin = fs * win
+        nwin = int(fs * win)
         # overlap = nwin - int(shift * fs)
 
         #####################################
@@ -1323,7 +1324,7 @@ class Speech():
         # 1. Const.
         f_min = 0. # The minimal frequency of the filter bank
         f_max = fs / 2
-        nwin = fs * win
+        nwin = int(fs * win)
 
         #####################################
         # 2. Speech.
